requirements.txt
Django==4.2.6
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-cachalot==2.6.1
django-environ==0.11.2
django-filter==21.1
django-silk==5.0.4
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
django-heroku
asgiref==3.6.0
Django==4.2.6
dj-database-url==0.5.0
django-storages==1.12.3
boto3==1.24.12
botocore==1.27.12
